psycology/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.7.0
django-countries==7.6.1
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
phonenumbers==8.13.54
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
